Split userAuth tests
